Move version_changes to ListsHelper
