+++ spec/controllers/lists_controller_spec.rb
@@ -115,31 +115,4 @@ RSpec.describe ListsController, type: :controller do
     
   end
 
-  describe 'version changes' do
-    
-    before(:each) do 
-      @lc = ListsController.new
-      @changeset = {"is_ranked"=>[false, true], "name"=>['bob', 'alice']}
-    end
-
-    it 'turns a hash into a string' do
-      expect(@lc.version_changes(@changeset)).to be_a(String)
-    end
-    
-    it 'prettifies the hash'
-    
-  end
-
-  describe 'nil_string' do
-
-    it 'returns "nil" if given nil' do
-      expect(ListsController.nil_string(nil)).to eq("nil")
-    end
-    
-    it 'returns the obj if not given nil' do 
-      expect(ListsController.nil_string("something")).to eq("something")
-    end
-    
-  end
-
 end
